chore(test): speed up e2e cases
