+++ tests/log/trace.test.ts
@@ -6,11 +6,11 @@ import { runRstestCli } from '../scripts';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-describe('console trace', () => {
-  it('should console log trace when printConsoleTrace enabled', async () => {
+describe('console log', () => {
+  it('should not console log when onConsoleLog return false', async () => {
     const { cli } = await runRstestCli({
       command: 'rstest',
-      args: ['run', 'log.test', '--printConsoleTrace'],
+      args: ['run', 'log.test', '-c', 'consoleLogFalse.config.ts'],
       options: {
         nodeOptions: {
           cwd: join(__dirname, 'fixtures'),
@@ -21,21 +21,19 @@ describe('console trace', () => {
     await cli.exec;
     const logs = cli.stdout.split('\n').filter(Boolean);
 
-    expect(logs.filter((log) => log.startsWith('I'))).toMatchInlineSnapshot(`
-      [
-        "I'm log",
-        "I'm warn",
-        "I'm error",
-        "I'm info",
-      ]
-    `);
-    expect(logs.some((log) => log.includes('log.test.ts:4:11'))).toBeTruthy();
+    expect(logs.filter((log) => log.startsWith('I'))).toEqual([]);
   });
 
-  it('should console trace correctly', async () => {
+  it('should onConsoleLog will not take effect when disableConsoleIntercept', async () => {
     const { cli } = await runRstestCli({
       command: 'rstest',
-      args: ['run', 'trace.test'],
+      args: [
+        'run',
+        'log.test',
+        '-c',
+        'consoleLogFalse.config.ts',
+        '--disableConsoleIntercept',
+      ],
       options: {
         nodeOptions: {
           cwd: join(__dirname, 'fixtures'),
@@ -46,6 +44,7 @@ describe('console trace', () => {
     await cli.exec;
     const logs = cli.stdout.split('\n').filter(Boolean);
 
-    expect(logs.some((log) => log.includes('trace.test.ts:4:11'))).toBeTruthy();
+    expect(logs.some((log) => log.startsWith('I'))).toBeTruthy();
+    expect(logs.some((log) => log.includes('log.test.ts:4:11'))).toBeFalsy();
   });
 });
